+++ src/main/java/com/chateat/chatEAT/config/SecurityConfig.java
@@ -59,7 +59,7 @@ public class SecurityConfig {
                 .with(new CustomFilterConfigurer(), Customizer.withDefaults())
                 .authorizeHttpRequests(authorizeRequests -> authorizeRequests
                         .requestMatchers("/auth/login", "/auth/reissue", "/members/**", "/swagger-ui/**",
-                                "/api-docs/**")
+                                "/api-docs/**", "/chat")
                         .permitAll()
                         .anyRequest().authenticated())
                 .logout(logout -> logout.deleteCookies("JSESSIONID"));
