+++ src/test/java/com/chateat/chatEAT/api/ChatControllerTest.java
@@ -38,7 +38,7 @@ public class ChatControllerTest {
     @Autowired
     private MockMvc mockMvc;
 
-    @Autowired
+    @MockBean
     private ChatRepository chatRepository;
 
     @MockBean
@@ -72,21 +72,5 @@ public class ChatControllerTest {
         // 응답 확인
         String responseContent = result.getResponse().getContentAsString();
         assertThat(responseContent).isEqualTo("Mock AI Response");
-    }
-
-    @DisplayName("Mongodb 상호작용에 성공한다.")
-    @Test
-    public void testSaveChatToMongoDB() {
-        // Given
-        Chat chat = new Chat();
-        chat.setMessage("Hello, MongoDB!");
-        chat.setTimestamp(LocalDateTime.now());
-
-        // When
-        Chat savedChat = chatRepository.save(chat);
-
-        // Then
-        assertThat(savedChat.getId()).isNotNull(); // 저장된 후 ID가 생성되었는지 확인
-        assertThat(chatRepository.findById(savedChat.getId())).isPresent(); // 데이터가 저장되었는지 확인
     }
 }
